back_fastapi/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.68
botocore==1.35.68
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.17
rsa==4.9
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
